+++ Src/utils.h
@@ -11,7 +11,7 @@ QString UIntToVersion(quint64 version_int, bool full = false);
 bool FilterVersion(QStringList& versions, QString version);
 QStringList FindFiles(QString dir, QStringList criteria);
 QStringList FindDirs(QString dir, QStringList criteria);
-QString BytesToString(uint32_t bytes);
+QString BytesToString(quint64 bytes);
 bool CopyFolder(QString input_dir, QString output_dir, std::function<void(int,int,QString)> callback);
 
 enum DIRECTORY_TYPE
